Add Code to interploate colors for LEDs
Animate the LEDs, both the DotStar and the NeoPixel all at once.

Add some debug code to test the Color interpolation and conversion
between HSV and RGB that runs on a posix style system.  Compiles and
runs on my mac in ITerm2; Probably works in Linux in Konsole or st.